backend/requirements.txt
# Core
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-dotenv==1.0.0

# Data & ML
pandas==2.1.4
numpy==1.24.3
yfinance==0.2.33
ta==0.11.0
scikit-learn==1.3.2

# TensorFlow (CPU version for speed)
tensorflow-cpu==2.15.0

# Database
pymongo==4.6.1
dnspython==2.4.2

# Utilities
requests==2.31.0
python-engineio==4.8.0
python-socketio==5.10.0
gunicorn==21.2.0